+++ software/src/RTC.h
@@ -10,7 +10,9 @@
 class RTC {
     public:
         RTC(int cePin, int ioPin, int clkPin);
-        void setTime(uint8_t year, uint8_t month, uint8_t day_of_month,                      uint8_t hour, uint8_t minute, uint8_t second,                    uint8_t day_of_week);
+        void setTime(uint16_t year, uint8_t month,  uint8_t day_of_month,
+                      uint8_t hour, uint8_t minute, uint8_t second,
+                      uint8_t day_of_week);
 
         void setTime(Time t);
         String printTime();
